+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+# Version 0.1.5
+- Add SimpleIoReactor::new_boxed.
+- AutoSendBuffer impl std::io::Write instead of std::fmt::Write
+- TODO: support non-socket reactors.
+
 # Version 0.1.4
 - Two stage AutoSendBuffer: write messages to send_buf and auto send on drop.
-- TODO: support non-socket reactors.
 
 # Version 0.1.3
 - Support Result<_, String>. return Err to close reactor.